Clarify role of user and course services
User service is only responsible for fetching and storing user
data. It does not cache any of the data itself, and consuming
sources must serialize data to be saved when saving, and
deserialize when loading from it.

In this case, the course service must take the serialized
CourseInfo objects and convert them into Courses.